Overhauled Buttons and Navigation regarding stock navigation bar
All settings for the stock navigation bar (hardware and on-screen) are now
located on the Navigation tab. Also, the ability to turn the on-screen navbar
on and off is now hidden if the device does not have hardware navbar buttons
(it's now always assumed that the on-screen navbar is enabled).