Updated power menu settings for Android 11
Settings to toggle the global actions panel and device settings on the power
menu were added to Android 11, so we now include them here (reworked to fit
with our settings code).

In addition, the power menu settings menu was moved from Buttons to User
Interface, as it made more sense to have it there.